refactor: new package mutex-manger
by Jacob
