+++ packages/core/server/package.json
@@ -25,10 +25,7 @@
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
     "semver": "^7.3.7",
-    "xpipe": "^1.0.5",
-    "ioredis": "^5.3.2",
-    "redlock": "^5.0.0-beta.2",
-    "async-mutex": "^0.3.2"
+    "xpipe": "^1.0.5"
   },
   "devDependencies": {
     "@types/semver": "^7.3.9"
